Modification of nac-license API for file download
