+++ api/views/__init__.py
@@ -1,10 +1,7 @@
-import requests
-
-from django.shortcuts import render, redirect
+from django.shortcuts import render
 from rest_framework.response import Response
 from rest_framework import viewsets, permissions, status
 from rest_framework.filters import SearchFilter, OrderingFilter
-from rest_framework.views import APIView
 
 from ..serializers import ProgSerializer, CalSerializer
 from prog.serializers import UnlockSerializer, UnlockUpdateSerializer, RaspeediSerializer
@@ -105,23 +102,6 @@ class UnlockViewSet(viewsets.ModelViewSet):
             return UnlockUpdateSerializer
         else:
             return UnlockSerializer
-
-
-class NacLicenseView(APIView):
-    authentication_classes = (TokenAuthSupportQueryString,)
-    permission_classes = (permissions.IsAuthenticated,)
-
-    def get(self, request, format=None):
-        if request.GET.get('update') and request.GET.get('uin'):
-            url = "https://majestic-web.mpsa.com/mjf00-web/rest/LicenseDownload"
-            payload = {
-                "mediaVersion": request.GET.get('update'),
-                "uin": request.GET.get('uin')
-            }
-            response = requests.get(url, params=payload, allow_redirects=True)
-            if response.status_code == 200:
-                return redirect(response.url)
-        return Response({"error": "Request failed"}, status=404)
 
 
 class DefaultCodeViewSet(viewsets.ModelViewSet):
